backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
python-multipart
pypdf2
python-dotenv
psycopg2-binary
langchain
langchain-openai
openai
pytest
passlib